requirements.txt
demucs
ffmpeg-python
pydub